Fix network install
The authorized_key files were not being created in the network boot
case, presumably because the cp /cdrom/... being used for the iso
installer case would not work when the install files were being fetched
from tftp. Storing the ssh key in the preseed file as text made it
possible to avoid copying any files from the install media in the late
actions, avoiding this problem (and possibly others).

The dropbear-initramfs service also was not seeing the configuration
file changes from the late action (although this worked fine for the iso
installer); adding a manual call to update-initramfs in the late actions
fixed this.